Normative: Remove line-cutting semantics for Atomics.wait (#20)
The line-cutting semantics was originally motivated by the anomaly that
resolving Atomics.waitAsync Promises does no meaningful work if the
agent is blocked by an Atomics.wait.

However, it's hard to reason about line-cutting semantics because it
undermines the currently simple FIFO fairness model, _and_ it also
doesn't actually solve the problem. The core ergonomics problem is the
futility of resolving Atomics.waitAsync Promises while the agent is
blocked -- regardless of location. The Atomics.wait/wake semantics,
however, is staunchly per-location. Solving the very narrow blocked
agent problem for a particular location is not worth the complexity.